# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Tesla Pro Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';
import type { OrderType } from './orderType';

export interface Order {
  id: string;
  userId: string;
  userEmail: string;
  userName: string;
  type: OrderType;
  description: string;
  amount: number;
  status: OrderStatus;
  createdAt: Date;
}