# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Tesla Pro Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}