# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Tesla Pro Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsOverview {
  totalUsers: number;
  newUsersThisWeek: number;
  totalRevenue: number;
  totalInvested: number;
  pendingTransactions: number;
  activeInvestments: number;
  totalOrders: number;
  pendingOrders: number;
}